etl/experimental/adaptive-sampling-client/backend/VWE_WorldDataAPI/pyproject.toml
[project]
name = "vwe-world-data-api"
version = "1.0.0"
description = "World data processing API for VWE adaptive sampling"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.2",
    "numpy>=2.1.2",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.7.1",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

